Client: Reorganizing components into new sub directories.
