+++ client/src/App.js
@@ -4,8 +4,8 @@ import { Switch, Route, Redirect } from 'react-router-dom';
 import UserWrapper from "./components/UserWrapper";
 import HeaderMenu from "./components/HeaderMenu";
 import Welcome from "./components/Welcome";
-import Create from "./components/Create";
-import Movies from "./components/Movies";
+import Create from "./components/create/Create";
+import Movies from "./components/movies/Movies";
 import Login from "./components/Login";
 import Logout from "./components/Logout";
 import Footer from "./components/Footer";
